tao lop kem giao vien day,phonghoc, thoikhoabieu;fix lay thong tin lop kem giaovien

+++ src/main/java/com/example/E4Center/models/PhongHoc.java
@@ -15,7 +15,7 @@ public class PhongHoc {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer maphong;
+    private Long maphong;
 
     @Column(length = 255)
     private String tenphong;
